moved tool under Analysis/law
