+++ Analysis/law/law.cfg
@@ -1,6 +1,7 @@
 [modules]
 
-analysis.tasks
+ShuffleMergeSpectral.tasks
+Hadd.tasks
 
 
 [job]
