+++ Analysis/law/setup.sh
